Fixed bugs and added printing results to file

+++ src/main/java/pl/edu/mimuw/bajttrade/adaptery/ZasobyJson.java
@@ -1,13 +1,15 @@
 package pl.edu.mimuw.bajttrade.adaptery;
 
+import com.squareup.moshi.ToJson;
+
 public class ZasobyJson {
-  public int diamenty;
+  public double diamenty;
   public int ubrania;
   public int narzedzia;
   public int jedzenie;
   public int programy;
 
-  public ZasobyJson(int diamenty, int ubrania, int narzedzia, int jedzenie, int programy) {
+  public ZasobyJson(double diamenty, int ubrania, int narzedzia, int jedzenie, int programy) {
     this.diamenty = diamenty;
     this.ubrania = ubrania;
     this.narzedzia = narzedzia;
